add list of crosses in frontend
